#1.1 fix: firebaseからのデータ取得タイミングの変更　他
