+++ components/organisms/modal/HeaderModal.tsx
@@ -4,7 +4,6 @@ import { useRecoilState, useSetRecoilState } from 'recoil'
 import { doc, DocumentReference, DocumentSnapshot, getDoc, onSnapshot, QueryDocumentSnapshot, setDoc } from 'firebase/firestore';
 
 import { styled } from '@mui/system';
-import PersonIcon from '@mui/icons-material/Person';
 import SettingsIcon from '@mui/icons-material/Settings';
 import LogoutIcon from '@mui/icons-material/Logout';
 
@@ -12,7 +11,6 @@ import { auth, db } from '../../../firestore/firebase';
 import { profileState } from '../../../state/profileState'
 import { headerModalShowState } from '../../../state/headerModalShowState';
 import { logoutModalShowState } from '../../../state/logoutModalShowState';
-import { selectPickUpPhotoState } from '../../../state/selectPickUpPhotoState';
 import { ProfileType } from '../../../types/ProfileType';
 import styles from '../../../styles/header.module.css'
 
@@ -28,8 +26,7 @@ const MyLogoutIcon = styled(LogoutIcon)({
 
 const HeaderModal = () => {
 
-  // mypageの仮選択のピックアップフォトの値をuseRecoilで管理
-  const setSelectPickUpPhoto = useSetRecoilState(selectPickUpPhotoState);
+
 
   // ヘッダーのモーダルの値をrecoilで管理
   const [headerModalShow, setHeaderModalShow] = useRecoilState(headerModalShowState);
@@ -37,46 +34,6 @@ const HeaderModal = () => {
   const setLogoutModalShow = useSetRecoilState(logoutModalShowState);
   // プロフィール情報をuseRecoilで管理
   const [profile, setProfile] = useRecoilState(profileState);
-
-  // 初回レンダリング時にprofile情報をfirebaseから取得してrecoilにセットする。
-  // firebaseにprofile情報がない場合は、初期値を設定し、firebaseに送信してrecoilにもセットする。
-  useEffect(()  => {
-    const initialRendering = async() => {
-      if(!auth.currentUser) return;
-      // firebaseからユーザー情報を取得
-      const initialProfile = doc(db, "users", auth.currentUser.uid) as DocumentReference<ProfileType>;
-      const snapProfile =  await getDoc<ProfileType>(initialProfile);
-      // 取得に成功した場合
-      if (snapProfile.exists()){
-        // firebaseのprofile情報をrecoilにセット
-        await setProfile(snapProfile.data());
-        await setSelectPickUpPhoto(snapProfile.data().userPickUpPhoto);
-      // 取得できなかった(データが入っていなかった)場合
-      } else {
-        // profileの初期値を設定
-        const profileExapmle = {
-          userName: "No Name",
-          userMemo: "よろしくお願いします。",
-          userPickUpPhoto: "https://firebasestorage.googleapis.com/v0/b/v-photo.appspot.com/o/image%2Fno_image.png?alt=media&token=bf785fef-6b0a-4640-9a22-4be838549468",
-          userPickUpDescription: "",
-          userIcon: "https://firebasestorage.googleapis.com/v0/b/v-photo.appspot.com/o/image%2Ficon3.png?alt=media&token=433d387c-2bcb-4f25-9593-a05b5d4dcb84",
-        }
-        // 初期値を設定する関数
-        const initialProfileValue = async() => {
-          if(auth.currentUser === null) return;
-          // firebaseにprofile情報の初期値を送信
-          const docRef = doc(db, "users", auth.currentUser.uid)
-          await setDoc(docRef, profileExapmle);
-          // recoilにも初期値をセット
-          await setProfile(profileExapmle);
-          await setSelectPickUpPhoto(profileExapmle.userPickUpPhoto)
-        }
-        initialProfileValue()
-      }
-    }
-    initialRendering()
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  },[]);
 
   // レンダリング時に走る関数
   useEffect(() => {
